refactor : CouponItemCardList 컴포넌트 삭제하고 CouponContentSection에서 처리
