+++ src/components/couponContentSection/CouponContentSection.styled.ts
@@ -6,3 +6,11 @@ export const StyledCouponContentSection = styled.div`
   gap: 4px;
   width: 100%;
 `;
+
+export const StyledCouponItemCardList = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+
+  width: 100%;
+`;
