--- src/components/couponItemCardList/CouponItemCardList.styled.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const StyledCouponItemCardList = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-
-  width: 100%;
-`;
